requirements.txt
opensearch-py>=2.5.0
PyYAML>=6.0.1
streamlit>=1.40.0
pytest>=8.0.0
